requirements.txt
selenium==3.141.0
pyzmq>=20.0.0
python-dotenv==0.15.0
httpx==0.16.1
redis==3.5.3
pid==3.0.4
pandas==1.2.1